rename inject to bean
