+++ src/main/java/com/patryklikus/winter/beans/BeanProviderImpl.java
@@ -1,6 +1,6 @@
-package com.patryklikus.winter.injection;
+package com.patryklikus.winter.beans;
 
-import com.patryklikus.winter.injection.Bean.Bean;
+import com.patryklikus.winter.beans.Bean.Bean;
 import com.patryklikus.winter.lifecycle.Close;
 import com.patryklikus.winter.lifecycle.Init;
 import com.patryklikus.winter.lifecycle.Run;
