+++ src/main/java/com/patryklikus/winter/beans/Beans.java
@@ -1,4 +1,4 @@
-package com.patryklikus.winter.injection;
+package com.patryklikus.winter.beans;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
