+++ src/test/java/beans/exampleProject/Main.java
@@ -1,4 +1,4 @@
-package injection.exampleProject;
+package beans.exampleProject;
 
 public class Main {
     public static void main(String[] args) {
