+++ src/test/java/beans/exampleProject/models/fruits/Apple.java
@@ -1,0 +1,4 @@
+package beans.exampleProject.models.fruits;
+
+public class Apple implements Fruit {
+}
