+++ src/test/java/beans/exampleProject/models/fruits/Banana.java
@@ -1,0 +1,4 @@
+package beans.exampleProject.models.fruits;
+
+public class Banana implements Fruit {
+}
