+++ src/test/java/beans/exampleProject/models/fruits/Fruit.java
@@ -1,0 +1,4 @@
+package beans.exampleProject.models.fruits;
+
+public interface Fruit {
+}
