--- src/test/java/injection/exampleProject/models/fruits/Apple.java
@@ -1,4 +1,0 @@
-package injection.exampleProject.models.fruits;
-
-public class Apple implements Fruit {
-}
