--- src/test/java/injection/exampleProject/models/fruits/Banana.java
@@ -1,4 +1,0 @@
-package injection.exampleProject.models.fruits;
-
-public class Banana implements Fruit {
-}
