--- src/test/java/injection/exampleProject/models/fruits/Fruit.java
@@ -1,4 +1,0 @@
-package injection.exampleProject.models.fruits;
-
-public interface Fruit {
-}
